Small enhancement for dynamically added forms
In case forms are added after the validator plugin is loaded, the `$(window).load()` will miss those. To pick up those I added a `$(document).on()` event handler installing the validator the first time a form is used. There might be a small performance hit of doing that check with each form, but this is very very unlikely it makes any difference in your app.