fixing options persistence at new calls
the problem was that options were not been saved, because at merge of options (default, current, new) the current options were not considered as were lost on the process, "$this.data()" didn't work at least for this purpose.

the second problem was that the constructor of new validator was depend of "if(!data)" that was only true on the first call, what caused to call with new options were never saved at data.